Complete porting to google's protobuf

+++ types/enums/enums.hpp
@@ -9,19 +9,19 @@
 
 namespace srilakshmikanthanp::clipbirdesk::types::enums {
 /// @brief IP type used in the discovery packet
-enum class IPType : quint8 {
+enum class IPType : quint32 {
   IPv4 = 0x00,
   IPv6 = 0x01,
 };
 
 /// @brief Allowed Error Codes
-enum ErrorCode : quint8 {
+enum ErrorCode : quint32 {
   CodingError   = 0x01,
   InvalidPacket = 0x02,
 };
 
 /// @brief Allowed Authentication Status
-enum AuthStatus : quint8 {
+enum AuthStatus : quint32 {
   AuthOkay  = 0x00,
   AuthFail  = 0x01,
 };
